+++ llvm/tools/llvm-dwp/CMakeLists.txt
@@ -1,18 +1,11 @@
 set(LLVM_LINK_COMPONENTS
   AllTargetsAsmParsers
   AllTargetsCodeGens
-  AllTargetsDescs
-  AllTargetsInfos
-  AsmPrinter
-  DebugInfoDWARF
-  MC
-  Object
-  Support
-  Target
   )
 
 add_llvm_tool(llvm-dwp
   llvm-dwp.cpp
+  DWP.cpp
   DWPError.cpp
 
   DEPENDS
